Updates the get_url checksum for bash completion
Now pinning the version of the remote src via git hash. We're already
hardcoding the checksum, so it makes sense to fetch only a single
version of the bash completion script. Doing so does mean we'll need to
manually review and bump vars to take advantage of upstream changes.

Fixes #1.